LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaDevices-enumerateDevices-persistent-permission.https-expected.txt



FAIL enumerateDevices depends only on capture state, not permission state assert_equals: At most one of a kind prior to capture expected 2 but got 5
